data project/requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0
selenium==4.15.2
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.26.2

# NLP and Entity Recognition
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1

# Text Similarity and Deduplication
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# LLM API Integration
openai==1.3.7
anthropic==0.7.7

# HTML Generation
jinja2==3.1.2
markdown==3.5.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
colorama==0.4.6

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
